fixed webpack, file structure, updated 404 page
