+++ src/App.jsx
@@ -1,22 +1,21 @@
 import React, { useContext } from 'react';
-import { Header, Footer } from './components';
-import { Routes } from './config/Routes';
+import { BrowserRouter as Router } from 'react-router-dom';
+import { Header, Footer, Routes } from './components';
 import { DarkModeContext } from './contexts/darkMode';
 import { GlobalStyles } from './styles/globalStyles';
 import * as S from './App.styles';
 
 const App = () => {
   const { isDarkMode } = useContext(DarkModeContext);
-
   return (
-    <>
+    <Router>
       <Header />
       <GlobalStyles isDarkMode={isDarkMode} />
       <S.PageContainer>
         <Routes />
       </S.PageContainer>
       <Footer />
-    </>
+    </Router>
   );
 };
