+++ src/components/index.js
@@ -15,3 +15,4 @@ export { default as Input } from './Inputs/Input/Input';
 export { default as TextArea } from './Inputs/TextArea/TextArea';
 export { default as Rate } from './StarRating/StarRating';
 export { default as ThemeSwitcher } from './ThemeSwitcher/ThemeSwitcher';
+export { default as Routes } from './Routes/Routes';
