--- src/config/Routes/index.js
@@ -1,1 +1,0 @@
-export { default as Routes } from './Routes';
