+++ src/index.jsx
@@ -1,5 +1,4 @@
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router } from 'react-router-dom';
 import DarkModeProvider from './contexts/darkMode';
 import { Provider } from 'react-redux';
 import store from './store';
@@ -8,9 +7,7 @@ import App from './App.jsx';
 ReactDOM.render(
   <Provider store={store}>
     <DarkModeProvider>
-      <Router>
-        <App />
-      </Router>
+      <App />
     </DarkModeProvider>
   </Provider>,
   document.getElementById('app')
